% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visGPX.R
\name{visGPX}
\alias{visGPX}
\title{Visualize GPX files}
\usage{
visGPX(file, df = NULL, dummy, ...)
}
\arguments{
\item{file}{GPX filename}

\item{df}{Optional data.frame with columns lon,lat,ele (num), time (POSIXct)
and runtime (num, in seconds).
DEFAULT: NULL (generated from \code{file})}

\item{dummy}{Dummy}

\item{\dots}{Further arguments passed to \code{\link{plot}}}
}
\value{
ReturnValue
}
\description{
Interactive Map from OSMtracker GPS tracks
}
\examples{

file <- system.file("extdata/Dranse_2017-06-06.gpx", package="OSMscale")
visGPX(file)

}
\seealso{
\code{\link{help}}, \code{\link{help}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{aplot}
