% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visGPX.R
\name{visGPX}
\alias{visGPX}
\title{Visualize GPX files}
\usage{
visGPX(file, df = NULL, element = 1, threshold_na = NA, smooth = 9,
  plot_interactive = TRUE, provider = "OpenStreetMap", plot_static = TRUE,
  whiteline = TRUE, bgmap = NULL, mfrow = c(2, 2), keeppar = FALSE, ...)
}
\arguments{
\item{file}{GPX filename}

\item{df}{Optional data.frame with columns lon,lat,ele (num) + time (POSIXct)
DEFAULT: NULL (internally generated from \code{file})}

\item{element}{Numeric: Element in GPX tracks list to be used. DEFAULT: 1}

\item{threshold_na}{Values above this speed are set to NA for smoothing.
Use e.g. 15 (kmh) for implausible values when walking.
DEFAULT: NA (ignored)}

\item{smooth}{Smoothing width passed to \code{berryFunctions::\link{movAv}}
for speed column. DEFAULT: 5}

\item{plot_interactive}{Logical: plot leaflet map? DEFAULT: TRUE}

\item{provider}{Character string: \code{leaflet::\link[leaflet]{providers}}
to be used for background map, e.g. "Esri.WorldImagery".
DEFAULT: OpenStreetMap"}

\item{plot_static}{Logical: plot static graphs? DEFAULT: TRUE}

\item{whiteline}{Logical: white background line in interactive map? DEFAULT: TRUE}

\item{bgmap}{Static background map from \code{OSMscale::\link{pointsMap}}.
Plotting is done faster if provided. DEFAULT: NULL}

\item{mfrow}{Multiple figures \link{par}ameter. DEFAULT: c(2,2)}

\item{keeppar}{Logical: keep paramters of plot (to add to last panel /
outer margins)? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{OSMscale::\link{pointsMap}}}
}
\value{
ReturnValue
}
\description{
Interactive Map from OSMtracker GPS tracks
}
\examples{
file <- system.file("extdata/Dranse_2017-06-06.gpx", package="visGPX")
# file <- "../../Sonstiges/2017-06-10_PORT.gpx"
visGPX(file, threshold_na=13, plot_static=FALSE)

file <- system.file("extdata/rad.gpx", package="visGPX")
visGPX(file, plot_static=FALSE) # with waypoints

\dontrun{# Excluded from checks to reduce computing time (map download)
dranse <- visGPX(file, threshold_na=13)

# Exporting:
pdf("dranse.pdf", height=5)
visGPX(file, bgmap=dranse$bgmap) # speed up by re-using static map
dev.off()
htmlwidgets::saveWidget(dranse$map, "dranse.html")
}

}
\seealso{
\code{\link{help}}, \code{\link{help}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
\keyword{aplot}
